requirements.txt
Django>2.2,<=3.0.14 #checkout azure supported version
django-import-export
djangorestframework
djangorestframework-simplejwt
django-admin-list-filter-dropdown
django-rest-swagger
django-parler
django-parler-rest
django-modeladmin-reorder
django-bootstrap-datepicker-plus==3.0.5
django-smart-selects>=1.5.9
django-storages[azure]
django-admin-menu>=1.5
data-wizard>=1.3.0 #checkout latest data wizard
msal>=1.9.0 #library needed for OIDC
mysqlclient>=2.0.1
Pillow>=6.0.0
python-dotenv>=0.17.1
pytz>=2018.5
tablib[all]
whitenoise

# data science and form packages
fuzzywuzzy # string matching algorithm
python-Levenshtein # helper lib for fuzzywuz
pandas==1.4.3
matplotlib==3.5.3
numpy==1.23.2
django-pandas==0.6.6
seaborn==0.11.2
scipy==1.9.0
django-crispy-forms==1.14.0
SQLAlchemy<=1.4.40

